python3/gitreview/requirements-dev.txt
astroid==1.5.3
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.2
PyYAML==3.12
requests==2.18.4
six==1.10.0
urllib3==1.22
wrapt==1.10.11
